Remove OpenBLAS source code requirement
The only files required from the source were config.h and param.h.
config.h is now generated from openblas_config.h, which is taken from
the installation directory. param.h has been included in this project.
param.h is required to obtain the right unroll parameters M and N and
depends on the CPU in config.h and may change in future versions of
OpenBLAS. The safest thing would be to take it from the same source
used to generate the installation, but since the source could not be
always present and there is no standard place where to find it,
that only file has been included here to make life simpler. Getting it
with git from the latest version could be just as unsafe as keeping
the current version, as both could differ from the version used to
compile the installed version.